Add config a file
